// src/interfaces/funcionarios/registerEmployee.ts
export interface IEmployeeBody {
  nome:            string;
  funcao:          string;
  data_nascimento: string;
  cpf:             string;
  email:           string;
  cargo:           string;
  salario:         number;
  contabancaria:   string;
  dataentrada:     string;
  cep:             string;
  logradouro:      string;
  bairro:          string;
  cidade:          string;
  estado:          string;
  numero:          string;
  complemento:     string;
  numerotelefone:  string;
  numerosetor:     string;
}